==> Pre-processing/3_soil_data_cleaning.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import warnings

df = pd.read_csv('../Data/raw/soil/soil_dataset.csv')

df

soil_state_unique = df['STATE'].unique()

df2 = pd.read_excel('../Data/clean/crop_yield.xlsx')

crop_state_unique = df2['State'].unique()

# elements that are present in soil_state_unique but not in crop_state_unique
set(soil_state_unique) - set(crop_state_unique)

# elements that are present in crop_state_unique but not in soil_state_unique
set(crop_state_unique) - set(soil_state_unique)

# Aggregating soil features per state
- rainfall feature not considered, because we already have data on it

states = []
ph = []
humidity = []
temperature = []
n_soil = []
p_soil = []
k_soil = []

n_soil_avg = df.groupby(['STATE'])['N_SOIL'].mean().reset_index()
p_soil_avg = df.groupby(['STATE'])['P_SOIL'].mean().reset_index()
k_soil_avg = df.groupby(['STATE'])['K_SOIL'].mean().reset_index()
temp_avg = df.groupby(['STATE'])['TEMPERATURE'].mean().reset_index()
humidity_avg = df.groupby(['STATE'])['HUMIDITY'].mean().reset_index()
ph_avg = df.groupby(['STATE'])['ph'].mean().reset_index()

average_soil_df = pd.DataFrame({
    'STATE': n_soil_avg['STATE'],
    'Mean_N_SOIL': n_soil_avg['N_SOIL'],
    'Mean_P_SOIL': p_soil_avg['P_SOIL'],
    'Mean_K_SOIL': k_soil_avg['K_SOIL'],
    'Mean_TEMPERATURE': temp_avg['TEMPERATURE'],
    'Mean_HUMIDITY': humidity_avg['HUMIDITY'],
    'Mean_PH': ph_avg['ph'],
})

average_soil_df

For states in our final merged crop data, that does not have their soil data, we have narrowed them down to 4.

For these 4 states, we assume that the soil conditions are similar to that of neighbouring states closest to it.
1. Bihar (closest state -> West Bengal)
2. Dadra and Nagar Haveli (closest state -> Gujarat)
3. Jharkhand (closest state -> West Bengal)
4. Mizoram (closest state -> Mizoram)

warnings.filterwarnings("ignore")

bihar_row = average_soil_df[average_soil_df['STATE'] == 'West Bengal']
bihar_row['STATE'] = 'Bihar'

dadra_row = average_soil_df[average_soil_df['STATE'] == 'Gujarat']
dadra_row['STATE'] = 'Dadra and Nagar Haveli'

jharkand_row = average_soil_df[average_soil_df['STATE'] == 'West Bengal']
jharkand_row['STATE'] = 'Jharkhand'

mizoram_row = average_soil_df[average_soil_df['STATE'] == 'Tripura']
mizoram_row['STATE'] = 'Mizoram'

average_soil_df = pd.concat([average_soil_df, bihar_row])
average_soil_df = pd.concat([average_soil_df, dadra_row])
average_soil_df = pd.concat([average_soil_df, jharkand_row])
average_soil_df = pd.concat([average_soil_df, mizoram_row])

# average_soil_df
average_soil_df

Furthermore, there are spelling differences between 4 states that overlaps for crop data and soil data.

Thus, we will be changing the spelling of these 4 states to match the name of the states in crop data for easier merging of data.

1. Andaman and Nicobar (changing to Andaman and Nicobar Islands)
2. Chattisgarh (Changing to Chhattisgarh)
3. Pondicherry (Changing to Puducherry)
4. Uttrakhand (Changing to Uttarakhand)

df['STATE'].unique()

df2['State'].unique()

### Renaming mistyped state names in soil data

average_soil_df['STATE'] =average_soil_df['STATE'].replace('Andaman and Nicobar', 'Andaman and Nicobar Islands')
average_soil_df['STATE'] =average_soil_df['STATE'].replace('Chattisgarh', 'Chhattisgarh')
average_soil_df['STATE'] =average_soil_df['STATE'].replace('Pondicherry', 'Puducherry')
average_soil_df['STATE'] =average_soil_df['STATE'].replace('Uttrakhand', 'Uttarakhand')

average_soil_df.to_csv('../Data/clean/soil_data.csv', index=False)
